feat(build): deploy mathy_pydoc package to pypi
 - trim out the mkdocs config integration because it didn't fit my use-case and it looked complex.
