+++ .github/workflows/main.yml
@@ -39,3 +39,9 @@ jobs:
           PYPI_USERNAME: ${{ secrets.PYPI_USERNAME }}
           PYPI_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
         run: cd libraries/mathy_alpha_sm && sh tools/deploy.sh
+      - name: Deploy mathy_pydoc
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          PYPI_USERNAME: ${{ secrets.PYPI_USERNAME }}
+          PYPI_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
+        run: cd libraries/mathy_pydoc && sh tools/deploy.sh
