--- libraries/mathy_pydoc_markdown/.versionupgrade
@@ -1,6 +1,0 @@
-# https://github.com/NiklasRosenstein/py-nr
-# $ nr versionupgrade {VERSION}
-
-tag v{VERSION}
-upgrade pydocmd/__init__.py:__version__ = '{VERSION}'
-upgrade setup.py:  version='{VERSION}',
