+++ libraries/mathy_pydoc_markdown/mathy_pydoc/about.py
@@ -1,0 +1,7 @@
+__title__ = "mathy_pydoc"
+__version__ = "0.6.5"
+__summary__ = "Minimal python API documentation generation for https://mathy.ai"
+__uri__ = "https://mathy.ai"
+__author__ = "Justin DuJardin"
+__email__ = "justin@dujardinconsulting.com"
+__license__ = "MIT"
