+++ libraries/mathy_pydoc_markdown/mathy_pydoc/document.py
