+++ libraries/mathy_pydoc_markdown/tests/test_markdown_headers.py
@@ -1,7 +1,7 @@
 import pytest
 import io
 
-from pydocmd.document import Section
+from mathy_pydoc.document import Section
 
 
 @pytest.fixture
