+++ libraries/mathy_pydoc_markdown/tests/test_restructuredtext.py
@@ -1,7 +1,7 @@
 import pytest
 
-from pydocmd.document import Section
-from pydocmd.restructuredtext import Preprocessor
+from mathy_pydoc.document import Section
+from mathy_pydoc.restructuredtext import Preprocessor
 
 
 @pytest.fixture
