+++ package.json
@@ -88,6 +88,7 @@
             "libraries/mathy_alpha_sm/mathy_alpha_sm/model.config.json",
             "libraries/website/docs/changelog.md",
             "libraries/mathy_python/mathy/about.py",
+            "libraries/mathy_python/mathy_pydoc_markdown/about.py",
             "package.json"
           ],
           "message": "chore(release): ${nextRelease.version} \n\n${nextRelease.notes}"
